Moved the splitting of the spritesheets into BufferedImageArrays into the BufferedImageLoader-class
The Splitting of the Spritesheets into BufferedImageArrays was moved
from the entity-class's constructor into the BufferedImageLoader. The
BufferedImageLoader now has added utility which also better fits into
this class instead of the entity-class. The new methods are called by
the Textures-class. The needed parameters can be set in this class and
the resulting BufferedImageArrays are now stored in it. Not only is it
cleaner to handle and store the textures in the Textures class, now
there is only one Array per Spritesheet stored instead of one per
entity. The Entities now only get references to these Arrays.